+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `scalr_iam_team`: fixed `users` attribute behaviour when not set in configuration ([#309](https://github.com/Scalr/terraform-provider-scalr/pull/309))
+
 ## [1.9.0] - 2024-02-23
 
 ### Added