+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `scalr_iam_team`: fixed `users` attribute behaviour when not set in configuration ([#309](https://github.com/Scalr/terraform-provider-scalr/pull/309))
 
+### Changed
+
+- `scalr_agent_pool`: attribute `environment_id` is deprecated ([#311](https://github.com/Scalr/terraform-provider-scalr/pull/311))
+
 ## [1.9.0] - 2024-02-23
 
 ### Added